Put CSSDiff back into gulpfile
